DTO-148 Fix: ERD 변경에 따른 데일리 생성 API 수정
